+++ apps/builder/src/features/auth/components/OnboardingPage.tsx
@@ -8,7 +8,7 @@ import {
   chakra,
   useColorModeValue,
 } from '@chakra-ui/react'
-import { Standard } from '@typebot.io/nextjs'
+import { Standard } from '@flowdacity/nextjs'
 import { useRouter } from 'next/router'
 import { useEffect, useRef, useState } from 'react'
 import confetti from 'canvas-confetti'