+++ apps/builder/src/features/publish/components/embeds/modals/Javascript/JavascriptBubbleSnippet.tsx
@@ -7,7 +7,7 @@ import {
 } from '../../snippetParsers'
 import { useTypebot } from '@/features/editor/providers/TypebotProvider'
 import { CodeEditor } from '@/components/inputs/CodeEditor'
-import { BubbleProps } from '@typebot.io/nextjs'
+import { BubbleProps } from '@flowdacity/nextjs'
 
 type Props = Pick<BubbleProps, 'theme' | 'previewMessage'>