+++ apps/builder/src/features/publish/components/embeds/modals/React/ReactPopupSnippet.tsx
@@ -1,6 +1,6 @@
 import { CodeEditor } from '@/components/inputs/CodeEditor'
 import { useTypebot } from '@/features/editor/providers/TypebotProvider'
-import { PopupProps } from '@typebot.io/nextjs'
+import { PopupProps } from '@flowdacity/nextjs'
 import parserBabel from 'prettier/parser-babel'
 import prettier from 'prettier/standalone'
 import { parseReactPopupProps } from '../../snippetParsers'
@@ -11,7 +11,7 @@ export const ReactPopupSnippet = ({
   const { typebot } = useTypebot()
 
   const snippet = prettier.format(
-    `import { Popup } from "@typebot.io/react";
+    `import { Popup } from "@flowdacity/react";
 
       const App = () => {
         return <Popup ${parseReactPopupProps({