+++ apps/builder/src/features/publish/components/embeds/settings/PopupSettings.tsx
@@ -7,7 +7,7 @@ import {
   HStack,
   Text,
 } from '@chakra-ui/react'
-import { PopupProps } from '@typebot.io/nextjs'
+import { PopupProps } from '@flowdacity/nextjs'
 import { useState, useEffect } from 'react'
 import { isDefined } from '@typebot.io/lib'