+++ apps/builder/src/features/publish/components/embeds/snippetParsers/shared.ts
@@ -1,4 +1,4 @@
-import { BotProps } from '@typebot.io/nextjs'
+import { BotProps } from '@flowdacity/nextjs'
 import parserBabel from 'prettier/parser-babel'
 import prettier from 'prettier/standalone'
 import { isDefined } from '@typebot.io/lib'
@@ -43,8 +43,8 @@ export const parseReactBotProps = ({ typebot, apiHost }: BotProps) => {
 }
 
 export const typebotImportCode = isCloudProdInstance()
-  ? `import Typebot from 'https://cdn.jsdelivr.net/npm/@typebot.io/js@0.2/dist/web.js'`
-  : `import Typebot from 'https://cdn.jsdelivr.net/npm/@typebot.io/js@${packageJson.version}/dist/web.js'`
+  ? `import Typebot from 'https://cdn.jsdelivr.net/npm/@flowdacity/js@0.2/dist/web.js'`
+  : `import Typebot from 'https://cdn.jsdelivr.net/npm/@flowdacity/js@${packageJson.version}/dist/web.js'`
 
 export const parseInlineScript = (script: string) =>
   prettier.format(