+++ apps/builder/src/features/settings/components/SettingsPage.tsx
@@ -1,6 +1,6 @@
 import { Seo } from '@/components/Seo'
 import { Flex } from '@chakra-ui/react'
-import { Standard } from '@typebot.io/nextjs'
+import { Standard } from '@flowdacity/nextjs'
 import { SettingsSideMenu } from './SettingsSideMenu'
 import { TypebotHeader } from '@/features/editor/components/TypebotHeader'
 import { useTypebot } from '@/features/editor/providers/TypebotProvider'