+++ packages/embeds/wordpress/trunk/includes/class-flowdacity-loader.php
@@ -1,6 +1,6 @@
 <?php
 
-class Typebot_Loader
+class Flowdacity_Loader
 {
 	protected $actions;
 	protected $filters;